refactor(plugin-sdk): rename plain text tool-call compat wrapper
