+++ scripts/lib/plugin-sdk-entrypoints.json
@@ -278,7 +278,6 @@
   "provider-model-shared",
   "provider-onboard",
   "provider-stream-family",
-  "provider-stream-runtime-internal",
   "provider-stream-shared",
   "provider-transport-runtime",
   "provider-stream",
