--- src/plugin-sdk/provider-stream-runtime-internal.ts
@@ -1,5 +1,0 @@
-/**
- * @deprecated Use `createPlainTextToolCallPromotionWrapper` from
- * `openclaw/plugin-sdk/provider-stream-shared`.
- */
-export { createPlainTextToolCallPromotionWrapper } from "./provider-stream-shared.js";
