+++ src/plugin-sdk/provider-stream.test.ts
@@ -4,14 +4,14 @@ import { VERSION } from "../version.js";
 import {
   composeProviderStreamWrappers as composeProviderStreamWrappersShared,
   createMoonshotThinkingWrapper as createMoonshotThinkingWrapperShared,
-  createPlainTextToolCallPromotionWrapper as createPlainTextToolCallPromotionWrapperShared,
+  createPlainTextToolCallCompatWrapper as createPlainTextToolCallCompatWrapperShared,
   createToolStreamWrapper as createToolStreamWrapperShared,
 } from "./provider-stream-shared.js";
 import {
   buildProviderStreamFamilyHooks,
   composeProviderStreamWrappers,
   createMoonshotThinkingWrapper,
-  createPlainTextToolCallPromotionWrapper,
+  createPlainTextToolCallCompatWrapper,
   createToolStreamWrapper,
   GOOGLE_THINKING_STREAM_HOOKS,
   KILOCODE_THINKING_STREAM_HOOKS,
@@ -67,9 +67,7 @@ describe("composeProviderStreamWrappers", () => {
 
   it("re-exports shared helper wrappers", () => {
     expect(createMoonshotThinkingWrapper).toBe(createMoonshotThinkingWrapperShared);
-    expect(createPlainTextToolCallPromotionWrapper).toBe(
-      createPlainTextToolCallPromotionWrapperShared,
-    );
+    expect(createPlainTextToolCallCompatWrapper).toBe(createPlainTextToolCallCompatWrapperShared);
     expect(createToolStreamWrapper).toBe(createToolStreamWrapperShared);
   });
